\name{ddeploy}
\alias{ddeploy}
\alias{ddeploy-package}
\docType{package}
\title{
ddeploy: Wrapper for the Duke Deploy REST API
}
\description{
ddeploy provides the functionality to deploy any predictive model to the cloud using the Duke Analytics RESTful API service. See <http://www.dukeanalytics.com/products/dukedeployR.php> for more details.
}
\details{ 
\tabular{ll}{
Package: \tab ddeploy\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab 2015-05-04\cr
License: \tab GPL\cr
}

}
\author{
Niall McGearailt <niall@dukeanalytics.com>
}
\references{
http://www.dukeanalytics.com/products/dukedeployR.php
}

\keyword{ model }
\keyword{ deploy }
\keyword{ deployment }
\keyword{ cloud }

