% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/duke_deploy.R
\name{duke_deploy}
\alias{duke_deploy}
\title{Deploy model using REST API}
\usage{
duke_deploy(auth_details, model_object, predict_udf = NULL,
  host_name = NULL)
}
\arguments{
\item{auth_details}{authorization parameters (must be stored in list format: list(user_name='exampleusername',api_key='exampleapikey')).}

\item{model_object}{the model to deploy e.g. lm() or glm() object.}

\item{predict_udf}{user defined function as alternative to the built-in model object predict function.}

\item{host_name}{optional parameter.}
}
\value{
None
}
\description{
Deploy model using REST API
}
\examples{
duke_auth <- list(
user_name = "try_it",
api_key   = "db1542b66f16aba5768d8a19c27dec4facf9168a",
endpoint = "/api/v1.0"
)
example_data <- as.data.frame(cbind(gl(3,50),rnorm(150)));names(example_data) <- c("x","y")
example_fit  <- lm(y~x,data=example_data)
duke_deploy(auth_details=duke_auth,model_object=example_fit)
}

