% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/duke_list.R
\name{duke_list}
\alias{duke_list}
\title{List all models deployed.}
\usage{
duke_list(auth_details)
}
\arguments{
\item{auth_details}{authorization parameters (must be stored in list format: list(user_name='exampleusername',api_key='exampleapikey')).}
}
\value{
A data.frame object with columns model_name, model_size, last_update, deploy_type.
}
\description{
List all models deployed.
}
\examples{
duke_auth <- list(
user_name = "try_it",
api_key   = "db1542b66f16aba5768d8a19c27dec4facf9168a",
endpoint = "/api/v1.0"
)
duke_list(auth_details=duke_auth)
}
\seealso{
\code{\link{duke_deploy}} to deploy a model or \code{\link{duke_predict}} for making predictions with deployed models.
}

