% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/duke_list.R
\name{duke_list}
\alias{duke_list}
\title{List deployed models}
\usage{
duke_list(auth_details)
}
\arguments{
\item{auth_details}{authorization parameters (must be stored in list format: list(user_name='exampleusername',api_key='exampleapikey')).}
}
\value{
None
}
\description{
List deployed models
}
\examples{
duke_auth <- list(
user_name = "try_it",
api_key   = "db1542b66f16aba5768d8a19c27dec4facf9168a",
endpoint = "/api/v1.0"
)
duke_list(auth_details=duke_auth)
}

